Initial commit: Added translation project for .resx files
This commit includes the initial version of the project that translates .resx files from English to other languages using Azure Text Translation service. The project is set up to be easily configurable for different source files and target languages. This is the first step in making our software more accessible to users worldwide.